Update identity provider schema tests
